multi-agent-team/requirements.txt
autogen>=0.2.0
openai>=1.0.0
jsonschema>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
typing-extensions>=4.0.0 